requirements.txt
asgiref==3.5.2
Django==3.2.12
django-ckeditor==6.5.0
django-js-asset==2.0.0
fontawesomefree==6.1.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
